requirements.txt
appdirs==1.4.3
certifi==2017.4.17
click==6.7
Flask==0.12.2
future==0.16.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-telegram-bot==6.0.1
six==1.10.0
transitions==0.5.2
Werkzeug==0.12.2
